Add y axis label to co2 graph
This commit adds a y axis label to the CO2 graph. While the unit is
printed above the chart with the current number (and the page title)
it's not always clear what units the graph is. To make it a bit clearer
this commit adds a yaxis to the graph.