remove blockers from setup
This change tries to eliminate any stumbling blocks for new users. It
specifies all the prerequisits needed before getting started, and
tries to make sure no critical steps are skipped. An
overrides.yaml.sample is provided.

In addition all `bx` commands are changed to `ibmcloud`, and new
output that shows the ingress controller configuration is listed.